liang

+++ src/main/java/com/yangchenle/electricityconsumptionsystem/dto/CopyRecordDTO.java
@@ -3,6 +3,7 @@ package com.yangchenle.electricityconsumptionsystem.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -13,7 +14,7 @@ public class CopyRecordDTO implements Serializable{
 
     private Integer electricId;
 
-    private Double copyData;
+    private BigDecimal copyData;
 
     private Date addTime;
