+++ src/main/java/com/yangchenle/electricityconsumptionsystem/entity/CopyRecordEntity.java
@@ -2,6 +2,7 @@ package com.yangchenle.electricityconsumptionsystem.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -11,7 +12,7 @@ public class CopyRecordEntity {
 
     private Integer electricId;
 
-    private Double copyData;
+    private BigDecimal copyData;
 
     private Date addTime;
